feat: implement specification and statistic controllers and routes, and update category routing in app.js
